HW2_Digit-Recognition-Problem/requirements.txt
matplotlib==3.10.1
numpy==2.2.4
opencv_python==4.11.0.86
opencv_python_headless==4.11.0.86
Pillow==11.2.1
torchmetrics==1.7.0